chore(web-finance): apply finance target in terraform
skip creation of cloud run service because it will need another
manual deployment step. The docker image will be built at that point,
but deploying the image cant be done before the initial infrastructure
is created. therefor anything that can be built up to the point that the
docker image can be deployed will be deployed here, then the image will
be pushed manually, and after the image was successfully pushed to the
freshly created infrastructure, the cloud run service will be deployed
by terraform using that manually deployed image.